Implemented as hard constraint
