+++ src/georgstage/controller.py
@@ -1,7 +1,8 @@
 from datetime import date, timedelta
 from dateutil.parser import parse
 import logging
-from georgstage.autofill import AutoFiller
+from georgstage.autofill import lp
+# from georgstage.autofill import greedy
 
 from georgstage.model import GeorgStage, Vagt, Opgave
 from georgstage.view import View
@@ -12,7 +13,7 @@ logger = logging.getLogger()
 class Controller(object):
     """docstring for Controller."""
 
-    def __init__(self, autofiller=AutoFiller()):
+    def __init__(self, autofiller=lp.autofill):
         model = GeorgStage([])
         self.model = model
         self.view = View(self, model)
@@ -184,7 +185,7 @@ class Controller(object):
                         raise ValueError(f'Skifte {skifte} not one of 1, 2, 3')
                     skifter[i] = skifte
             logger.info(f'Skifter: {skifter}')
-            fill_result = self.autofiller.autofill(self.model, skifter)
+            fill_result = self.autofiller(self.model, skifter)
             logger.info(fill_result)
             if fill_result.status != 1:
                 raise ValueError('Tjek at vagtplan er korrekt udfyldt')
